Search the airplane
airplaneData is responsible for searching the related fields in the database based on the unique ID of each airplane. Then, in ticketsData.ts, using the unique flightDetails of each airplane ticket, we will be able to determine which airplane it is. Currently, the name of the airplane will be used to determine the seating arrangement and the status of these seats in the future.