To assign seats randomly
The function 'findAdjacentSeatGroup' searches for a set of seats that are relatively close together for the number of passengers within a group. It takes into account that if one passenger is a minor, they should be seated next to an adult. For this purpose, passengers are separated based on their age.