Database connection control
I added the logic to establish the connection to the database using Prisma in your Node.js application with Express. I implemented a time control and automatic reconnection in case of connection failures. Now, when starting the application, it will attempt to establish the connection, and if it fails, reconnection attempts will be made every 5 seconds until a successful connection is established.